refactor: using transferTypes object to determine how different transfers should be processed
